Separate graphviz debugging and have BundleGraph subclass Graph directly
