+++ packages/core/core/src/Parcel.js
@@ -119,8 +119,15 @@ export default class Parcel {
   async build() {
     try {
       // console.log('Starting build'); // eslint-disable-line no-console
+      // $FlowFixMe This reliably fails on Windows. Not sure why.
       let assetGraph = await this.assetGraphBuilder.build();
-      // await graph.dumpGraphViz();
+
+      if (process.env.PARCEL_DUMP_GRAPH != null) {
+        const dumpGraphToGraphViz = require('@parcel/utils/src/dumpGraphToGraphViz')
+          .default;
+        await dumpGraphToGraphViz(assetGraph, 'MainAssetGraph');
+      }
+
       let bundleGraph = await this.bundle(assetGraph);
       await this.package(bundleGraph);
